상품 목록 불러오기 작업 중간 지점

+++ src/main/java/com/joo/abysshop/dto/product/ProductItemListResponseDTO.java
@@ -10,8 +10,8 @@ import org.springframework.web.multipart.MultipartFile;
 @Builder
 public class ProductItemListResponseDTO {
 
-    private Integer productId;
+    private Long productId;
     private String productName;
-    private Integer price;
+    private Long price;
     private MultipartFile image;
 }
